Canvas can be a static page
